+++ src/components/ContentHeader/index.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import './ContentHeader.scss';
+
+const ContentHeader = ({ children, isVideo }) => (
+  <div className={isVideo ? 'content__header--video' : 'content__header'}>
+    {children}
+  </div>
+);
+
+ContentHeader.propTypes = {
+  children: PropTypes.node
+};
+
+export default ContentHeader;