+++ src/components/Items/Album.js
@@ -1,25 +1,23 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
 
-import { queryString } from '../../helpers';
-import { ListLink } from '../Links';
 import Inline from '../Inline';
+import { ListLink, ArtistLink } from '../Links';
 
-export const AlbumItem = ({ value }) => (
+const Album = ({ value, kind }) => (
   <div className="item--album">
-    <section className="primary--album">
+    <div className="primary--album">
       <img
         className="img--album"
         src={value.artworkUrl100.replace('100x100', '200x200')}
         alt=""
       />
-    </section>
+    </div>
 
-    <section className="secondary--album">
+    <div className="secondary--album">
       <Inline>
         <ListLink
-          list="list__link--album"
+          list="album"
           name={value.collectionName}
           id={value.collectionId}
           explicit={value.collectionExplicitness}
@@ -27,17 +25,18 @@ export const AlbumItem = ({ value }) => (
         />
       </Inline>
 
-      <Link
-        className="link list__link--artist"
-        to={`/artist/${queryString(value.artistName)}/${value.artistId}`}
-      >
-        {value.artistName}
-      </Link>
-    </section>
+      {kind !== 'artist' ? (
+        <ArtistLink value={value} type="list" list={true} />
+      ) : (
+        <span style={{ fontSize: 13 }}>
+          {value.releaseDate.substring(0, 4)}
+        </span>
+      )}
+    </div>
   </div>
 );
 
-AlbumItem.propTypes = {
+Album.propTypes = {
   value: PropTypes.shape({
     artworkUrl100: PropTypes.string.isRequired,
     collectionName: PropTypes.string.isRequired,
@@ -47,3 +46,5 @@ AlbumItem.propTypes = {
     artistId: PropTypes.string.isRequired
   })
 };
+
+export default Album;