+++ src/components/Items/Video.js
@@ -1,12 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
 
-import { queryString } from '../../helpers';
-import { ListLink } from '../Links';
 import Inline from '../Inline';
+import { ListLink, ArtistLink } from '../Links';
 
-export const VideoItem = ({ value, contentList }) => (
+const Video = ({ value, contentList, kind }) => (
   <div className="item--video">
     <section className="primary--video">
       <img
@@ -25,7 +23,7 @@ export const VideoItem = ({ value, contentList }) => (
         {contentList ? <p className="index">{value.trackNumber}.</p> : null}
 
         <ListLink
-          list="list__link--video"
+          list="video"
           name={value.trackName}
           id={value.trackId}
           explicit={value.trackExplicitness}
@@ -34,18 +32,19 @@ export const VideoItem = ({ value, contentList }) => (
       </Inline>
 
       {!contentList ? (
-        <Link
-          className="link list__link--artist"
-          to={`/artist/${queryString(value.artistName)}/${value.artistId}`}
-        >
-          {value.artistName}
-        </Link>
+        kind !== 'artist' ? (
+          <ArtistLink value={value} type="list" list={true} />
+        ) : (
+          <span style={{ fontSize: 13 }}>
+            {value.releaseDate.substring(0, 4)}
+          </span>
+        )
       ) : null}
     </section>
   </div>
 );
 
-VideoItem.propTypes = {
+Video.propTypes = {
   value: PropTypes.shape({
     artworkUrl100: PropTypes.string.isRequired,
     trackName: PropTypes.string.isRequired,
@@ -55,3 +54,5 @@ VideoItem.propTypes = {
     artistId: PropTypes.string.isRequired
   })
 };
+
+export default Video;