+++ src/components/Players/index.js
@@ -1,0 +1,4 @@
+import Audio from './Audio';
+import Video from './Video';
+
+export { Audio as AudioPlayer, Video as VideoPlayer };