+++ src/components/Title/index.js
@@ -1,12 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Inline from './Inline';
+import Inline from '../Inline';
+import Explicit from '../Explicit';
+
+import './Title.scss';
 
 const Title = ({ title, explicit }) => (
   <Inline>
     <h2 className="title">{title}</h2>
-    <span className={explicit} />
+    <Explicit className={explicit} />
   </Inline>
 );