+++ src/containers/Artist.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 
-import Spinner from '../components/Spinner';
+import Preloader from '../components/Preloader';
+import HeaderTitle from '../components/HeaderTitle';
 import { ArtistList } from '../components/Lists';
 
 import { artist, list } from '../api';
@@ -13,57 +14,52 @@ export default class Artist extends Component {
 
     artist(id).then(data => this.setState({ artist: data.results[0] }));
 
-    list({ id: id, entity: 'song', limit: 15 }).then(data =>
+    list({ id, entity: 'song', limit: 15 }).then(data =>
       this.setState({ songs: data.results })
     );
 
-    list({ id: id, entity: 'album', limit: 20 }).then(data =>
+    list({ id, entity: 'album', limit: 20 }).then(data =>
       this.setState({ albums: data.results })
     );
 
-    list({ id: id, entity: 'musicVideo', limit: 12 }).then(data =>
+    list({ id, entity: 'musicVideo', limit: 12 }).then(data =>
       this.setState({ videos: data.results })
     );
   }
 
   render() {
+    const { artist, songs, albums, videos } = this.state;
+
     return (
       <>
-        <div className="header__title">
-          <h2>{this.state.artist ? this.state.artist.artistName : null}</h2>
-        </div>
+        <HeaderTitle>
+          <h2>{artist ? artist.artistName : null}</h2>
+        </HeaderTitle>
 
-        {this.state.songs ? (
-          <ArtistList
-            {...this.props}
-            values={this.state.songs}
-            className="scroller--songs"
-            type="Songs"
-          />
+        {songs ? (
+          <ArtistList values={songs} className="scroller--songs" type="Songs" />
         ) : (
-          <Spinner />
+          <Preloader />
         )}
 
-        {this.state.albums ? (
+        {albums ? (
           <ArtistList
-            {...this.props}
-            values={this.state.albums}
+            values={albums}
             className="scroller--albums"
             type="Albums"
           />
         ) : (
-          <Spinner />
+          <Preloader />
         )}
 
-        {this.state.videos ? (
+        {videos ? (
           <ArtistList
-            {...this.props}
-            values={this.state.videos}
+            values={videos}
             className="scroller--videos"
             type="Music videos"
           />
         ) : (
-          <Spinner />
+          <Preloader />
         )}
       </>
     );