+++ src/containers/Video.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 
-import Spinner from '../components/Spinner';
+import Preloader from '../components/Preloader';
+import View from '../components/View';
 import { VideoContent } from '../components/Contents';
 
 import { video } from '../api';
@@ -11,14 +12,20 @@ export default class Video extends Component {
   componentDidMount() {
     const { id } = this.props.match.params;
 
-    video(id).then(data => this.setState({ video: data.results[0] }));
+    video(id).then(data => {
+      const video = data.results[0];
+
+      this.setState({ video });
+    });
   }
 
   render() {
     return this.state.video ? (
-      <VideoContent value={this.state.video} />
+      <View className="video">
+        <VideoContent value={this.state.video} />
+      </View>
     ) : (
-      <Spinner />
+      <Preloader />
     );
   }
 }